finished with hybenert crud

+++ src/main/java/com/example/JspStudentCrud/DB/StudentDaoHbnt.java
@@ -78,7 +78,7 @@ public class StudentDaoHbnt {
      * @param id
      * @return
      */
-    public Student getStudent(int id) {
+    public Student getStudent(Long id) {
         Transaction transaction = null;
         Student student = null;
         try (Session session = HibernateUtil.getSessionFactory().openSession()) {
